Added fast inverse sqrt function. can now normalise vectors using fastinvsqrt to improve speed.
