+++ ballEngineWindows.cbp
@@ -132,10 +132,10 @@
 		<Unit filename="source/Events/classSceneManager.cpp" />
 		<Unit filename="source/Events/classSurvivalScene.cpp" />
 		<Unit filename="source/Interfaces/characterWorldInterface.cpp" />
+		<Unit filename="source/Miscellaneous/Math.cpp" />
 		<Unit filename="source/Miscellaneous/Observer.cpp" />
 		<Unit filename="source/Miscellaneous/class2DMatrix.cpp" />
 		<Unit filename="source/Miscellaneous/jsonParsing.cpp" />
-		<Unit filename="source/Miscellaneous/sfVectorMath.cpp" />
 		<Unit filename="source/Miscellaneous/stringConversion.cpp" />
 		<Unit filename="source/Miscellaneous/structs.cpp" />
 		<Unit filename="source/Renderer/classLight.cpp" />
